[REF] mail: move rtc related fields to call folder
Part of task-3605717
