+++ addons/mail/static/src/core/common/store_service.js
@@ -66,8 +66,6 @@ export class Store extends BaseStore {
     Notification;
     /** @type {typeof import("@mail/core/common/persona_model").Persona} */
     Persona;
-    /** @type {typeof import("@mail/discuss/call/common/rtc_session_model").RtcSession} */
-    RtcSession;
     /** @type {typeof import("@mail/core/common/settings_model").Settings} */
     Settings;
     /** @type {typeof import("@mail/core/common/thread_model").Thread} */
