+++ addons/mail/static/src/discuss/call/common/store_service_patch.js
@@ -1,5 +1,6 @@
 import { Record } from "@mail/core/common/record";
 import { Store } from "@mail/core/common/store_service";
+import { RtcSession } from "@mail/discuss/call/common/rtc_session_model";
 
 import { patch } from "@web/core/utils/patch";
 
@@ -7,6 +8,8 @@ import { patch } from "@web/core/utils/patch";
 const StorePatch = {
     setup() {
         super.setup(...arguments);
+        /** @type {typeof import("@mail/discuss/call/common/rtc_session_model").RtcSession} */
+        this.RtcSession = RtcSession;
         this.rtc = Record.one("Rtc", {
             compute() {
                 return {};
